+++ client/scripts/db/congo.js
@@ -84,6 +84,7 @@ module.exports = {
   skipWmcRomCheck: true,
   features: [
     'securityPic',
+    'wpc95',
   ],
   initialise: {
     closedSwitches: [