+++ client/scripts/db/corv.js
@@ -87,6 +87,7 @@ module.exports = {
   skipWmcRomCheck: true,
   features: [
     'securityPic',
+    'wpcSecure',
   ],
   initialise: {
     closedSwitches: [