+++ client/scripts/db/cv.js
@@ -83,6 +83,7 @@ module.exports = {
   skipWmcRomCheck: true,
   features: [
     'securityPic',
+    'wpc95',
   ],
   initialise: {
     closedSwitches: [