+++ client/scripts/db/i500.js
@@ -73,6 +73,7 @@ module.exports = {
   skipWmcRomCheck: true,
   features: [
     'securityPic',
+    'wpcSecure',
   ],
   initialise: {
     closedSwitches: [