+++ client/scripts/lib/sound.js
@@ -1,90 +1,84 @@
 'use strict';
 
+import { Howl, Howler } from 'howler';
+
 export { AudioOutput };
 
-function AudioOutput(AudioContext, sampleSize) {
-  return new Sound(AudioContext, sampleSize);
+function AudioOutput(sprite) {
+  return new Sound(sprite);
 }
 
-const DEFAULT_SAMPLE_SIZE = 1024 * 1;
-//at least the from DAC
-const INPUT_SAMPLE_RATE_HZ = 8000;
-const OUTPUT_SAMPLE_RATE_HZ = 44100;
-const PLAYBACK_RATE = INPUT_SAMPLE_RATE_HZ / OUTPUT_SAMPLE_RATE_HZ;
-const MONO = 1;
+const STOP_ALL_SOUNDS_ID = 0;
 
 class Sound {
 
-  constructor(AudioContext) {
-    this.audioCtx = new AudioContext();
-    this.gainNode = this.audioCtx.createGain();
-
-    // configure nodes
-    this.gainNode.gain.value = 0.9;
-
-    // connect nodes
-    this.gainNode.connect(this.audioCtx.destination);
-
-    try {
-      this._prepareAudio();
-      setInterval(() => {
-        this._fillAudioBuffer();
-      }, 50);
-    } catch(error) {
-      console.log('AUDIO_INIT_FAILED', error.message);
-      this.noAudio = true;
-    }
+  constructor(sprite = []) {
+    this.sprite = sprite;
+    //TODO pass audio sample
+    this.sound = new Howl({
+      src: [
+        '',
+        // TODO add bong sound
+       // 'sound/result.mp3',
+      ],
+      sprite,
+      onplayerror: (error) => {
+        console.log('SOUND PLAYER ERROR', error.message);
+      },
+      onloaderror: (error) => {
+        //console.log('SOUND LOAD ERROR', error.message);
+      },
+      onload: () => {
+        console.log('SOUND LOADED');
+      },
+    });
+    this.activePlayId = [];
   }
 
-  _prepareAudio() {
-    this.ym2151buffer = this.audioCtx.createBuffer(2, DEFAULT_SAMPLE_SIZE, OUTPUT_SAMPLE_RATE_HZ);
-    this.buffer = this.audioCtx.createBuffer(MONO, DEFAULT_SAMPLE_SIZE, INPUT_SAMPLE_RATE_HZ);
-    this.bufferData = this.buffer.getChannelData(0);
-    this.audioBuffer = this.audioCtx.createBufferSource();
-    this.audioBuffer.playbackRate.value = PLAYBACK_RATE;
-    //TODO is this leaking?
-    this.audioBuffer.connect(this.gainNode);
-    this.count = 0;
+  callback(message = {}) {
+    switch (message.command) {
+      case 'PLAYSAMPLE': {
+        const id = message.id;
+        const key = 'snd' + id;
+        if (!this.sprite[key]) {
+          return console.log('SAMPLE ID NOT FOUND', id);
+        }
+        console.log('playbackId', id);
+        this.activePlayId[0] = this.sound.play(key);
+        break;
+      }
+      case 'MAINVOLUME':
+        const volume = (message.value / 31);
+        this.setVolume(volume);
+        break;
+
+      case 'STOPSOUND':
+        this.sound.stop();
+        break;
+
+      case 'CHANNELOFF':
+        console.log('CHANNELOFF_NOT_IMPLEMENTED YET', message.channel);
+        break;
+
+      default:
+        console.log('NOT_IMPLEMENTED YET', message.command);
+        break;
+
+    }
+
   }
 
-  writeAudioData(value) {
-    if (this.noAudio) {
-      return;
-    }
-    if (this.count < DEFAULT_SAMPLE_SIZE) {
-      // fill buffer with values form -1.0 and 1.0
-      this.bufferData[this.count++] = (value - 64) / 64;
-    } else {
-      // set the buffer in the AudioBufferSourceNode
-      this.audioBuffer.buffer = this.buffer;
-//      console.log('sound output', this.audioBuffer.buffer.duration);
-
-      // start the source playing
-      this.audioBuffer.start();
-      //TODO need to do a clever buffer switch here
-      this._prepareAudio();
-    }
+  playBootSound() {
+    //TODO
   }
 
   setVolume(floatVolume) {
-    if (this.noAudio) {
-      return;
-    }
-    this.gainNode.gain.value = floatVolume;
-  }
-
-  _fillAudioBuffer() {
-//    this.mixStereoFunction(this.ym2151buffer, this.ym2151buffer.length, 0);
-//    console.log('bufr',this.ym2151buffer);
-  }
-
-  setMixStereoFunction(mixStereoFunction) {
-    this.mixStereoFunction = mixStereoFunction;
-//    console.log('XXXmixStereoFunction', this.mixStereoFunction);
+    console.log('SET MAINVOLUME', floatVolume);
+    Howler.volume(floatVolume);
   }
 
   stop() {
-    //audioSourceNode.disconnect()">
+    this.sound.pause();
   }
 
 }