+++ test/integration/benchmark.js
@@ -29,9 +29,15 @@ function benchmarkWithCycleCount(tickSteps) {
 
   const loadRomFilesPromise = Promise.all([
     loadFile(romU06Path),
-    loadFile(romU14Path).catch((error) => { debug(error.message); }),
-    loadFile(romU15Path).catch((error) => { debug(error.message); }),
-    loadFile(romU18Path).catch((error) => { debug(error.message); }),
+    loadFile(romU14Path).catch((error) => {
+      debug(error.message);
+    }),
+    loadFile(romU15Path).catch((error) => {
+      debug(error.message);
+    }),
+    loadFile(romU18Path).catch((error) => {
+      debug(error.message);
+    }),
   ]);
 
   return loadRomFilesPromise
@@ -50,7 +56,7 @@ function benchmarkWithCycleCount(tickSteps) {
       const ticksExecuted = wpcSystem.executeCycle(CYCLE_COUNT, tickSteps);
       const durationMs = Date.now() - tsStart;
       const status = wpcSystem.getUiState();
-      console.error(`  ${tickSteps}\t\t${durationMs}\t\t${status.cpuState.missedIRQ}\t\t${status.cpuState.missedFIRQ}\t\t${status.cpuState.tickCount}`);
+      console.error(`  ${tickSteps}\t\t${durationMs}\t\t${status.cpuState.missedIRQ}\t\t${status.cpuState.missedFIRQ}\t\t${ticksExecuted}`);
     });
 }